scripts/requirements_api.txt
Flask>=2.0.0
flask-cors>=3.0.0
PyTango>=9.3.0
Pillow>=8.0.0
requests>=2.25.0

